+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Made forms translatable.
 - Fixed map resize on drawer close.
 - Fixed an issue that caused the like button not to update.
+- Added `fitBounds` method to map.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.